recsys-service/requirements.txt
scikit-learn
numpy
clickhouse_driver
psycopg2-binary
transformers
torch
grpcio
grpcio-tools